refactor(flow): extract upload logic into useFileUploads and finalize duplicate policy
- Add useFileUploads hook for batch handling and error propagation
- Keep exact name+ext as duplicate error; clearer messages
- Ensure reducer-based updates for uploading -> uploaded/failed
