+++ packages/giselle/src/react/flow/hooks/index.ts
@@ -1,6 +1,7 @@
 export * from "./use-add-node";
 export * from "./use-connection";
 export * from "./use-copy-node";
+export * from "./use-file-uploads";
 export * from "./use-node-update";
 export * from "./use-properties-panel";
 export * from "./use-workflow-designer";
